Rename query builder modules to match component name
