+++ resources/frontend_client/app/query_builder/Calendar.react.js
