+++ resources/frontend_client/app/query_builder/CardFavoriteButton.react.js
