+++ resources/frontend_client/app/query_builder/DataReferenceField.react.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import DataReferenceQueryButton from './data_reference_query_button.react';
+import DataReferenceQueryButton from './DataReferenceQueryButton.react';
 import Icon from "metabase/components/Icon.react";
 import Query from "metabase/lib/query";
 import inflection from 'inflection';
