+++ resources/frontend_client/app/query_builder/DataReferenceTable.react.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import Icon from "metabase/components/Icon.react";
-import DataReferenceQueryButton from './data_reference_query_button.react';
+import DataReferenceQueryButton from './DataReferenceQueryButton.react';
 import inflection from 'inflection';
 
 import Query from "metabase/lib/query";
