+++ resources/frontend_client/app/query_builder/FieldSelector.react.js
