+++ resources/frontend_client/app/query_builder/FieldWidget.react.js
@@ -1,7 +1,7 @@
 "use strict";
 
-import FieldSelector from "./field_selector.react";
-import FieldName from "./field_name.react";
+import FieldSelector from "./FieldSelector.react";
+import FieldName from "./FieldName.react";
 import Icon from "metabase/components/Icon.react";
 import Popover from "metabase/components/Popover.react";
