+++ resources/frontend_client/app/query_builder/FilterWidget.react.js
@@ -1,11 +1,11 @@
 'use strict';
 /*global _*/
 
-import Calendar from './calendar.react';
+import Calendar from './Calendar.react';
 import Icon from "metabase/components/Icon.react";
-import FieldName from './field_name.react';
-import FieldSelector from './field_selector.react';
-import SelectionModule from './selection_module.react';
+import FieldName from './FieldName.react';
+import FieldSelector from './FieldSelector.react';
+import SelectionModule from './SelectionModule.react';
 import Popover from "metabase/components/Popover.react";
 import ColumnarSelector from "metabase/components/ColumnarSelector.react";
