+++ resources/frontend_client/app/query_builder/IconBorder.react.js
