+++ resources/frontend_client/app/query_builder/QueryHeader.react.js
@@ -2,18 +2,18 @@
 /*global setTimeout, clearTimeout*/
 
 import ActionButton from 'metabase/components/ActionButton.react';
-import AddToDashboard from './add_to_dashboard.react';
-import AddToDashSelectDashModal from '../components/AddToDashSelectDashModal.react';
-import CardFavoriteButton from './card_favorite_button.react';
-import DeleteQuestionModal from '../components/DeleteQuestionModal.react';
+import AddToDashboard from './AddToDashboard.react';
+import AddToDashSelectDashModal from 'metabase/components/AddToDashSelectDashModal.react';
+import CardFavoriteButton from './CardFavoriteButton.react';
+import DeleteQuestionModal from 'metabase/components/DeleteQuestionModal.react';
 import Header from "metabase/components/Header.react";
 import HistoryModal from "metabase/components/HistoryModal.react";
 import Icon from "metabase/components/Icon.react";
 import Modal from "metabase/components/Modal.react";
 import ModalWithTrigger from "metabase/components/ModalWithTrigger.react";
-import QueryModeToggle from './query_mode_toggle.react';
-import QuestionSavedModal from '../components/QuestionSavedModal.react';
-import SaveQuestionModal from '../components/SaveQuestionModal.react';
+import QueryModeToggle from './QueryModeToggle.react';
+import QuestionSavedModal from 'metabase/components/QuestionSavedModal.react';
+import SaveQuestionModal from 'metabase/components/SaveQuestionModal.react';
 
 import inflection from "inflection";
 import cx from "classnames";
