+++ resources/frontend_client/app/query_builder/QueryVisualizationTable.react.js
