+++ resources/frontend_client/app/query_builder/SearchBar.react.js
