+++ resources/frontend_client/app/query_builder/VisualizationSettings.react.js
